feat(auth): implement email and magic code authentication flow
- Add EmailSentConfirmation component to display email verification message.
- Create MagicCodeInput component for users to input the received magic code.
- Develop SignInForm component to handle email sign-in and provider sign-in options.
- Introduce SignInProviderButton component for OAuth provider sign-in.
- Implement useSignInHandlers hook to manage authentication state and logic.
- Add error handling and loading states for better user experience.
